por hoy dejo, es la primera ves que configuro un swager, y me daba la insertidumbre si documneta solo,cualquiera sea la tecnologia, removi varias cosas y me quedaahacer uniforme los outbup de los controladores

+++ src/main/java/sube/interviews/mareoenvios/ProductRepository.java
@@ -2,6 +2,6 @@ package sube.interviews.mareoenvios;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 
 }
