+++ src/main/java/sube/interviews/mareoenvios/ShippingItemRepository.java
@@ -2,6 +2,6 @@ package sube.interviews.mareoenvios;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShippingItemRepository extends CrudRepository<ShippingItem, Long> {
+public interface ShippingItemRepository extends CrudRepository<ShippingItem, Integer> {
 
 }
