+++ src/main/java/sube/interviews/mareoenvios/ShippingRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShippingRepository extends CrudRepository<Shipping, Long> {
+public interface ShippingRepository extends CrudRepository<Shipping,Integer> {
 
 	
 }
